Restructure Objects
- Moved them out into seperate files in their own folder
- Simplified namespace name => Removing horizontal whitespace
